v0.4.0

+++ src/main/java/software/ulpgc/bigdata/matrices/matrix/StandardMatrix.java
@@ -1,5 +1,5 @@
 package software.ulpgc.bigdata.matrices.matrix;
 
 public interface StandardMatrix extends Matrix {
-    long[][] get();
+    Double[][] get();
 }
